requirements.txt
asgiref==3.8.1
boto3==1.38.15
botocore==1.38.15
Django==5.2.1
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
gunicorn==23.0.0
jmespath==1.0.1
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
